CLose to finished now, just going to sort out dealer vs customer

+++ Data/Domain/Car.cs
@@ -9,7 +9,7 @@ namespace CarSales.Data.Domain
 
 	public class Car
 	{
-		public int CarID { get; private set; }
+		public Int64 CarID { get; private set; }
 		public string Make { get; private set; }
 		public string Model { get; private set; }
 		public int Year { get; private set; }
